+++ test/selectors/index.ts
@@ -1,1 +1,2 @@
 export * from "./css.edge.cases.spec.playwright";
+export * from "./custom.testid.attribute.spec.playwright";